Refactored TagCloudBlockController to pull logic out into controller service
