+++ src/Foundation/Features/Blog/TagCloudBlock/TagCloudBlockController.cs
@@ -1,4 +1,5 @@
-﻿using EPiServer;
+﻿using System;
+using EPiServer;
 using EPiServer.Core;
 using EPiServer.DataAbstraction;
 using EPiServer.Framework.DataAnnotations;
@@ -19,43 +20,29 @@ namespace Foundation.Features.Blog.TagCloudBlock
         private readonly IContentLoader _contentLoader;
         private readonly BlogTagFactory _blogTagFactory;
         private readonly ICategoryContentLoader _categoryContentLoader;
+        private readonly TagCloudBlockControllerService _controllerService;
 
         public TagCloudBlockController(IContentLoader contentLoader,
             CategoryRepository categoryRepository,
             BlogTagRepository blogTagRepository,
-            BlogTagFactory blogTagFactory, ICategoryContentLoader categoryContentLoader)
+            BlogTagFactory blogTagFactory,
+            ICategoryContentLoader categoryContentLoader,
+            TagCloudBlockControllerService controllerService)
         {
             _contentLoader = contentLoader;
             _blogTagFactory = blogTagFactory;
             _categoryContentLoader = categoryContentLoader;
+            _controllerService = controllerService ?? throw new ArgumentNullException(nameof(controllerService));
         }
 
         public override ActionResult Index(Cms.Blocks.TagCloudBlock currentBlock)
         {
             var model = new TagCloudBlockModel(currentBlock)
             {
-                Tags = GetTags(currentBlock.BlogTagLinkPage)
+                Tags = _controllerService.GetTags(currentBlock.BlogTagLinkPage)
             };
 
             return PartialView(model);
-        }
-
-        public IEnumerable<BlogItemViewModel.TagItem> GetTags(ContentReference startTagLink)
-        {
-            var tags = new List<BlogItemViewModel.TagItem>();
-            foreach (var item in BlogTagRepository.Instance.LoadTags())
-            {
-                var cat = _categoryContentLoader.GetFirstBySegment<StandardCategory>(item.TagName); // Assumes tag name == url segment
-                var url = string.Empty;
-
-                if (startTagLink != null)
-                {
-                    url = _blogTagFactory.GetTagUrl(_contentLoader.Get<PageData>(startTagLink.ToPageReference()), cat.ContentLink);
-                }
-
-                tags.Add(new BlogItemViewModel.TagItem() { Count = item.Count, Title = item.DisplayName, Weight = item.Weight, Url = url });
-            }
-            return tags;
         }
     }
 }
