+++ src/Foundation/Foundation.csproj
@@ -857,6 +857,7 @@
     <Compile Include="Features\Blog\BlogListPage\BlogListPageController.cs" />
     <Compile Include="Features\Blog\BlogListPage\BlogListPageControllerService.cs" />
     <Compile Include="Features\Blog\TagCloudBlock\TagCloudBlockController.cs" />
+    <Compile Include="Features\Blog\TagCloudBlock\TagCloudBlockControllerService.cs" />
     <Compile Include="Features\CatalogContent\CatalogContentControllerBase.cs" />
     <Compile Include="Features\CatalogContent\Variation\VariationController.cs" />
     <Compile Include="Features\Category\StandardCategoryController.cs" />
